♻️refactor: 선제적 제안 패키지 리팩토링
