+++ src/main/java/com/project/backend/domain/suggestion/converter/SuggestionConverter.java
@@ -11,8 +11,8 @@ import com.project.backend.domain.suggestion.enums.Status;
 import com.project.backend.domain.suggestion.enums.SuggestionType;
 import com.project.backend.domain.suggestion.util.SuggestionKeyUtil;
 import com.project.backend.domain.suggestion.util.TargetKeyHashUtil;
-import com.project.backend.domain.suggestion.vo.RecurrenceSuggestionCandidate;
-import com.project.backend.domain.suggestion.vo.SuggestionCandidate;
+import com.project.backend.domain.suggestion.generation.candidate.RecurrenceSuggestionCandidate;
+import com.project.backend.domain.suggestion.generation.candidate.SuggestionCandidate;
 import com.project.backend.domain.todo.entity.Todo;
 import com.project.backend.domain.todo.entity.TodoRecurrenceGroup;
 import lombok.AccessLevel;
