+++ src/main/java/com/project/backend/domain/suggestion/generation/detector/vo/MonthlyHistory.java
@@ -1,4 +1,4 @@
-package com.project.backend.domain.suggestion.detector.vo;
+package com.project.backend.domain.suggestion.generation.detector.vo;
 
 import java.util.List;
 import java.util.Set;
