+++ src/main/java/com/project/backend/domain/suggestion/generation/detector/vo/pattern/PatternValue.java
@@ -1,4 +1,4 @@
-package com.project.backend.domain.suggestion.detector.vo.pattern;
+package com.project.backend.domain.suggestion.generation.detector.vo.pattern;
 
 public sealed interface PatternValue
         permits IntervalValue, WeeklySetValue, MonthlySetValue {
