+++ src/main/java/com/project/backend/domain/suggestion/generation/detector/vo/pattern/WeeklySetValue.java
@@ -1,4 +1,4 @@
-package com.project.backend.domain.suggestion.detector.vo.pattern;
+package com.project.backend.domain.suggestion.generation.detector.vo.pattern;
 
 import java.time.DayOfWeek;
 import java.util.Set;
